% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{autocomplete_textOutput}
\alias{autocomplete_textOutput}
\title{Text input with tooltip's from Google}
\usage{
autocomplete_textOutput(
  key = Sys.getenv("google_key"),
  label = "Set a direction",
  inputId,
  width = NULL,
  placeholder = NULL,
  use_shinywidgets = FALSE,
  icon = NULL,
  size = NULL
)
}
\arguments{
\item{key}{Key for google maps to use API data (string)}

\item{label}{Text input label (string)}

\item{inputId}{Input ID (string)}

\item{width}{Horizontal size of widget. Value from 1 to 12}

\item{placeholder}{text inside widget (string)}

\item{use_shinywidgets}{Boolean. If it is TRUE, it will return the widget from the shinyWidgets library}

\item{icon}{Fontawesome icon}

\item{size}{Icon size}
}
\description{
Text input with tooltip's from Google
}
\examples{
if (interactive()) {
library(shiny)
library(ggmap)

key <- "YOUR_API_KEY"

shinyApp(
  ui = fluidPage(
    autocomplete_textOutput(label = "Set a direction", inputId = "direction"),
    verbatimTextOutput("text")
  ),
  server = function(input, output){
    observeEvent(input$direction,{
      output$text <- renderText({
        register_google(key)
        dir <- geocode(location = input$direction)
        if (is.na(dir$lat[1]) == T){
          "No address found"
        }else{
          paste0("lat: ",dir$lat, ", lng: ", dir$lon)
        }
      })
    })
  }
)

}
}
